Fix uw.function.evaluate / eliminate evalf
Rewrite of the uw.function.evaluate code. Now the code will check for points inside the domain and use PETSc interpolation for those. Points outside the domain are computed using the rbf_interpolation code which is perfectly happy to extrapolate wildly. This prevents crashes we were seeing with NodalPointSwarms / Semi-Lagrangian computations (particularly in curved domains).

Rather than exposing the evalf as a separate function, we now supply an argument to the evaluate method. We might want to change the name to rbf or something intuitive / meaningful.

Also had to rewrite the meshVariable / sympy printing code for each of the two functions as there were massive problems with the way the caching worked. I hope this is now fixed but, anyway, very strange that we did not see that bug appearing before
