+++ src/underworld3/visualisation.py
@@ -189,7 +189,7 @@ def scalar_fn_to_pv_points(pv_mesh, uw_fn, dim=None, simplify=True):
             dim = 3
 
     coords = pv_mesh.points[:, 0:dim]
-    scalar_values = uw.function.evalf(uw_fn, coords)
+    scalar_values = uw.function.evaluate(uw_fn, coords, evalf=True)
 
     return scalar_values
 
@@ -210,7 +210,7 @@ def vector_fn_to_pv_points(pv_mesh, uw_fn, dim=None, simplify=True):
     coords = pv_mesh.points[:, 0:dim]
     vector_values = np.zeros_like(pv_mesh.points)
 
-    vector_values[:, 0:dim] = uw.function.evalf(uw_fn, coords)
+    vector_values[:, 0:dim] = uw.function.evaluate(uw_fn, coords, evalf=True)
 
     return vector_values
 
@@ -228,7 +228,7 @@ def vector_fn_to_pv_points(pv_mesh, uw_fn, dim=None, simplify=True):
 #     vector_values = np.zeros_like(coords)
 
 #     for i in range(0, dim):
-#         vector_values[:, i] = uw.function.evalf(uw_fn[i], coords)
+#         vector_values[:, i] = uw.function.evaluate(uw_fn[i], coords, evalf=True)
 
 #     return vector_values
